Summit SOM 8M Plus: Add M7 support
Added a new recipe (lrd-m7-demos) which copies Laird-built application binaries
to the resulting image.

Part of this support is the addition of a set of devicetree source files (and
resulting binaries) that specify the necessary changes for the M7 core. One
noteable deviation from the NXP devicetree sources for M7 support is that the
reserved memory regions have been relocated to lower DDR addresses to support
SOMs with lower RAM sizes (512MB, 1GB).

An additional change is to increase the max buffer size for the RPMsg I2C
driver. The default size of 16 bytes is too small to accommodate a number of
relayed commands for the PAC193x part (when using the M7 for audio playback,
I2C commands to the I2C3 bus are pushed to the M7 core which relays the data
between the A core and the target peripheral).

The u-boot-imx defconfig has also been updated to enable support for U-Boot
commands required to boot the M7 core:
* Enable CONFIG_CMD_MEMORY (QSPI)

The u-boot boot logic has been updated to boot the M7 core based on the name of
the devicetree config which is used. A new script has been added
(set-m7-boot-mode) which can be used to trigger this logic (the M7 core is
disabled by default).

Bug 20965