feat: Support HAB-based i.MX bootloader signing
This change introduces optional support for secure boot for the NXP i.MX 8M
Plus-based products via the HAB. This support is added as an appended shell
script to the 'deploy' task for the imx-mkimage recipe and is enabled by
configuring ":summit-secure-hab" as an override in the local.conf file.

OVERRIDES .= ":summit-secure-hab"

Bootloader signing supports two options: locally or remotely.

For local signing, it expected that the NXP CST (code signing tool) is
independently provided for use with the build process (due to licensing
restrictions). Also, all of the supporting files/keys/certificates must
be provided as well.

For remote signing, a set of variables must be defined, including the URL of the
target signing server's REST endpoint. The provided script utilizes Microsoft's
MSAL libraries to authenticate with the signing server endpoint and uses the
RFPROS_SIGNING_xxxx variables for configuration. The script then POSTs the
unsigned bootloader binary and relevant signing address/offset values to the
signing server and receives back the signed bootloader binary.

Lastly the shell script updates the relevant symlink to pull the signed
bootloader binary into the remaining build process tasks.

Also introduced here is a new distro explicitly for use with WBx3 which enables
access to the U-Boot command line interface (sets "bootdelay" to 3).

Refs: PROD-14416